requirements.txt
appium-python-client>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
pytest>=6.0.0